Add composition test and xml file
I'm currently in the process of adding composition functionality to
Hobbes and the NVL. This commit adds two files, composition.xml and
tests/composition. Composition.xml specificies a composed set of
enclaves and a job on those enclaves. Tests/composition is a harness
for those frameworks. The code these files exercise is in the Hobbes
shell.