Automate cray build and clean up pmi forwarding
This patch automates the building of the nvl infrastructure, including a custom MPI and xhpcg for cray systems. It includes a custom build.pl in the nvl tree which will use the custom cray kernel in Kevin's home directory. This patch also includes a cleanup of the static forwarding infrastructure which fixes warnings and uses automake to automatically generate the library and executables makefiles.